requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.3
django-factory-boy==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.4
executing==1.2.0
factory-boy==3.2.1
Faker==15.2.0
idna==3.4
inflection==0.5.1
ipdb==0.13.9
ipython==8.6.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
openapi-codec==1.3.2
packaging==21.3
parameterized==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
toml==0.10.2
traitlets==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
